+++ CHANGELOG.md
@@ -8,7 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- [`extract`] and [`merge`] should optionally inject provenance [#977]
 - Checking for empty strings in addition to missing ones in `missing_label.rq` [#1017]
+- Add "domain" and "range" support to export [#1061]
 
 ### Fixed
 - Fixed report serialization in JSON [#1016]
@@ -309,6 +311,7 @@ First official release of ROBOT!
 [`diff`]: http://robot.obolibrary.org/diff
 [`explain`]: http://robot.obolibrary.org/explain
 [`export`]: http://robot.obolibrary.org/export
+[`expand`]: http://robot.obolibrary.org/expand
 [`extract`]: http://robot.obolibrary.org/extract
 [`filter`]: http://robot.obolibrary.org/filter
 [`measure`]: http://robot.obolibrary.org/measure
@@ -324,10 +327,14 @@ First official release of ROBOT!
 [`template`]: http://robot.obolibrary.org/template
 [`validate`]: http://robot.obolibrary.org/validate
 
+[#1061]: https://github.com/ontodev/robot/issues/1061
+[#1026]: https://github.com/ontodev/robot/issues/1026
+[#1017]: https://github.com/ontodev/robot/issues/1017
 [#1016]: https://github.com/ontodev/robot/issues/1016
 [#1009]: https://github.com/ontodev/robot/issues/1009
 [#979]: https://github.com/ontodev/robot/pull/979
 [#978]: https://github.com/ontodev/robot/pull/978
+[#977]: https://github.com/ontodev/robot/pull/977
 [#973]: https://github.com/ontodev/robot/pull/973
 [#971]: https://github.com/ontodev/robot/pull/971
 [#964]: https://github.com/ontodev/robot/pull/964