+++ CHANGELOG.md
@@ -7,9 +7,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### Fixed
-- input IRI now takes catalog file into account [#1030]
-
 ### Added
 - [`extract`] and [`merge`] should optionally inject provenance [#977]
 - Checking for empty strings in addition to missing ones in `missing_label.rq` [#1017]
@@ -18,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fixed report serialization in JSON [#1016]
 - Fix missing labels in [`diff`] output. [#1026]
+- input IRI now takes catalog file into account [#1030]
 
 ## [1.9.0] - 2022-06-16