+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fixed report serialization in JSON [#1016]
 - Fix missing labels in [`diff`] output. [#1026]
+- input IRI now takes catalog file into account [#1030]
 
 ## [1.9.0] - 2022-06-16
 
@@ -328,6 +329,7 @@ First official release of ROBOT!
 [`validate`]: http://robot.obolibrary.org/validate
 
 [#1061]: https://github.com/ontodev/robot/issues/1061
+[#1030]: https://github.com/ontodev/robot/issues/1030
 [#1026]: https://github.com/ontodev/robot/issues/1026
 [#1017]: https://github.com/ontodev/robot/issues/1017
 [#1016]: https://github.com/ontodev/robot/issues/1016